Split rewritten modules to separate .cpp/.h files
